create BrandView component
